+++ src/Web/Application.php
@@ -54,6 +54,7 @@ class Application extends \yii\web\Application
 
         parent::bootstrap();
 
+        $this->addConfiguredRoutes();
         $this->setDefaultEmail();
     }